requirements.txt
blinker==1.9.0
click==8.2.1
contourpy==1.3.3
cycler==0.12.1
Flask==3.1.2
fonttools==4.60.0
greenlet==3.2.4
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.9
MarkupSafe==3.0.2
matplotlib==3.10.6
numpy==1.26.4
packaging==25.0
pandas==2.3.2
pillow==11.3.0
PyMySQL==1.1.2
pyparsing==3.2.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
scikit-learn==1.7.1
scikit-surprise==1.1.4
scipy==1.16.1
seaborn==0.13.2
six==1.17.0
SQLAlchemy==2.0.43
threadpoolctl==3.6.0
typing_extensions==4.15.0
tzdata==2025.2
Werkzeug==3.1.3
